ZMQ skeleton
This is just a small proof-of-concept of the ZMQ socket design.
First each client connects a Request socket to the server's Reply socket.  The intent is that clients submit
actions that _they_ take to the server via this socket and the server responds with the result of the command.
Next, each client connects a Subscribe socket to the server's Publish socket.  The intent is that whenever the
world's gamestate changes, it's published to everybody.

So, say I want to "move".   Client sends "I move" via Request socket to server.  Server responds "okay, you didn't hit a wall" to the Client via the Reply socket.  The server THEN sends "Hey, Client X moved to spot Y" to ALL clients.